Moved normalize array functionality to schemaAndTables where most GREGoR-specific data handling is located
